Migrate make_linear_operator to jax_util. Also improve docs
