+++ cryojax/experimental/__init__.py
@@ -1,10 +1,10 @@
-from ..simulator._api_utils import make_linear_operator as make_linear_operator
 from ..simulator._multislice import (
     AbstractMultisliceIntegrator as AbstractMultisliceIntegrator,
     FFTMultisliceIntegrator as FFTMultisliceIntegrator,
     MultisliceScatteringTheory as MultisliceScatteringTheory,
 )
 from ..simulator._solvent_2d import (
+    AbstractRandomSolvent2D as AbstractRandomSolvent2D,
     GRFSolvent2D as GRFSolvent2D,
     SolventMixturePower as SolventMixturePower,
 )
