+++ cryojax/jax_util/__init__.py
@@ -14,13 +14,12 @@ from ._filter_specs import make_filter_spec as make_filter_spec
 from ._grid_search import (
     AbstractGridSearchMethod as AbstractGridSearchMethod,
     MinimumSearchMethod as MinimumSearchMethod,
-    MinimumSolution as MinimumSolution,
-    MinimumState as MinimumState,
     run_grid_search as run_grid_search,
     tree_grid_shape as tree_grid_shape,
     tree_grid_take as tree_grid_take,
     tree_grid_unravel_index as tree_grid_unravel_index,
 )
+from ._linear_operator import make_linear_operator as make_linear_operator
 from ._pytree_transforms import (
     AbstractPyTreeTransform as AbstractPyTreeTransform,
     CustomTransform as CustomTransform,
