+++ cryojax/jax_util/_grid_search/__init__.py
@@ -12,6 +12,4 @@ from .search_loop import run_grid_search as run_grid_search
 from .search_method import (
     AbstractGridSearchMethod as AbstractGridSearchMethod,
     MinimumSearchMethod as MinimumSearchMethod,
-    MinimumSolution as MinimumSolution,
-    MinimumState as MinimumState,
 )
